support players with null "first name"
Removed the requirement for first_name to be non-null, and allowed
the FIDE ID-picker to update a name parameter to be null.

This needs to be supported as some players have no first name in
the FIDE database. E.g. Sarthak Bathla.

There is also an icu_name change that goes along with this.