Add comment explaining why test may fail
This test relies on the Rails hasher being configured the same as
it is in production, which is not the default standard hasher.

Since this is configured in secrets.yml which is not version controlled, the
test can't be fixed for everyone.

Maybe this should be configured somewhere version controlled, instead.