fix test for unchanged date of birth
The old test was wrong but appears to have been passing accidentally, from
behaviour in Rails 4 where updating a Date field on an ActiveRecord
object with a String set the new value to be a String. Now it's a Date.

Added test in player_spec to confirm this regression. This test should
not pass in Rails 4.